MLX perf: the M3 Max was 76% idle — parallel scene synthesis + GEMM distances
Measured (d=48, batch 2): GPU step 78 ms, but 248 ms of single-threaded
float64 numpy scene synthesis serialized before every step — GPU busy
24%. Fixes: (1) Prefetcher — ProcessPoolExecutor scene workers with
prefetch depth workers+2 (--workers, default 0 = legacy inline);
make_scene float32 (24 full-frame Gaussian exps were 76 ms/cycle in
f64). (2) d2 in GEMM form |t|^2+|p|^2-2t.p — same math, 3x less memory
traffic, 78 -> 64 ms/step. Result: ~90 ms/step measured UNDER GPU
contention (vs 326) — 3.6x, ~5x solo. Remaining headroom noted:
mx.compile (64 -> 50 ms measured), bf16, larger batch. Judge stays
exact numpy f64 — laws judge, losses train.

+++ nn/v1/synth.py
@@ -27,14 +27,15 @@ def make_scene(rng, side=512, blobs=24):
     (clips in every frame above base). Gaussian color blobs over a
     global gradient, occasional hard edge; a squared ramp keeps most
     area in the low stops, photograph-like."""
-    yy, xx = np.mgrid[0:side, 0:side] / side
+    yy, xx = (np.mgrid[0:side, 0:side] / side).astype(np.float32)
     rgb = np.stack([0.008 + 0.06 * xx ** 2, 0.008 + 0.06 * yy ** 2,
                     0.008 + 0.06 * (1 - xx) ** 2], axis=-1)
     for _ in range(blobs):
         cy, cx = rng.random(2)
         sig = 0.03 + 0.15 * rng.random()
-        amp = rng.random(3) * (0.8 * rng.random() ** 2)   # few bright, many dim
-        g = np.exp(-(((yy - cy) ** 2 + (xx - cx) ** 2) / (2 * sig ** 2)))
+        amp = (rng.random(3) * (0.8 * rng.random() ** 2)).astype(np.float32)
+        g = np.exp(-(((yy - np.float32(cy)) ** 2 + (xx - np.float32(cx)) ** 2)
+                     / np.float32(2 * sig ** 2)))
         rgb += g[..., None] * amp[None, None, :]
     if rng.random() < 0.5:                      # a hard edge into shadow
         cut = int(side * (0.25 + 0.5 * rng.random()))
